langchain_app/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# LangChain
langchain==0.1.0
langchain-community==0.0.10
langchain-experimental==0.0.47

# Ollama
ollama==0.1.7

# Utilities
python-dotenv==1.0.0
requests==2.31.0